Collin/cad events

+++ fetchSpace.js
@@ -37,32 +37,37 @@ function parseObjectData(objectData) {
 function checkVisibility(objectPosition) {
     // find events added previously
     // event must be on this date
-    return Event.getByDate(objectPosition.date)
-    .then(events => {
-        // event must be for this body
-        events.forEach(event => {
-            if (event.body_id == objectPosition.body.id) {
-                // delete events
-                return event.delete()
-                // where we had checked the visibility before
-            }
+    // if it is not asteroid
+    if (objectPosition.body.body_type != 3) {
+        return Event.getByDate(objectPosition.date)
+        .then(events => {
+            // event must be for this body
+            return Promise.all(events.map(event => {
+                if (event.body_id == objectPosition.body.id) {
+                    // delete events
+                    return event.delete()
+                    // where we had checked the visibility before
+                } else {
+                    return null
+                }
+            }).filter(event => event != null))
         })
-    })
-    .then(() => {
-        const body = objectPosition.body
-        const objectRa = objectPosition.ra
-        const objectDec = objectPosition.dec
-        const currentTime = new Date()
-        Body.getByName('Sun')
-        .then(sun => {
-            return sun[0].getPosition()
-            .then(sunPosition => {
-                if (sunPosition.ra - 90 > objectPosition.ra || sunPosition.ra + 90 < objectPosition.ra) {
-                    // get all users
-                    return User.getAll()
-                    .then(users => {
-                        return Promise.all(users.map(user => {
+        .then(() => {
+            const body = objectPosition.body
+            const objectRa = objectPosition.ra
+            const objectDec = objectPosition.dec
+            const currentTime = new Date() 
+            return User.getAll()
+            .then(users => {
+                Body.getByName('Sun')
+                .then(sun => {
+                    return sun.getPosition()
+                    .catch(getSunPos)
+                    .then(sunPosition => {
+                        if (sunPosition.ra - 90 > objectPosition.ra || sunPosition.ra + 90 < objectPosition.ra) {
+                            // get all users
                             // get their positions
+                            return Promise.all(users.map(user => {
                                 if (user.lat > 0) {
                                     if (objectDec > 90 - user.lat) {
                                         // object is always visible at night
@@ -84,13 +89,27 @@ function checkVisibility(objectPosition) {
                                             .then(user.addEvent)
                                     }
                                 }
-                            })
-                        )
+                            }))
+                        }
                     })
-                }   
-            })         
+                })         
+            })
         })
-    })
+    } else {
+        return null
+    }
+}
+
+function getSunPos() {
+    return axios.get(`http://www.strudel.org.uk/lookUP/json/?name=Sun`)
+        .then(parseResponse)
+        .then(parseObjectData)
+        // body.addLocationPoint returns an object with ra and dec
+        .then(parsedObjectData => {
+            return Body.getByName('Sun')
+            .then(sun => sun.addLocationPoint(parsedObjectData.ra.decimal, parsedObjectData.dec.decimal))
+        })
+        
 }
 
 module.exports = fetchSpace
